+++ frontend/src/components/FormInput/FormInput.type.ts
@@ -1,0 +1,4 @@
+export interface FormInputError {
+  isError: boolean;
+  message: string;
+}
